+++ src/tools/index.ts
@@ -1,0 +1,14 @@
+import { fetchUrlTool, fetchUrl } from './fetchUrl.js';
+import { fetchUrlsTool, fetchUrls } from './fetchUrls.js';
+
+// Export tool definitions
+export const tools = [
+  fetchUrlTool,
+  fetchUrlsTool
+];
+
+// Export tool implementations
+export const toolHandlers = {
+  [fetchUrlTool.name]: fetchUrl,
+  [fetchUrlsTool.name]: fetchUrls
+};
